Add unit-tests system with empty tests for forum.js
There are used three frameworks:

* mocha -- http://mochajs.org/
  (functions describe(), it() and so on for TDD)

* chai  -- http://chaijs.com/
  (different assert functions for tests)

* sinon -- http://sinonjs.org/
  (mocks and stubs for tests)

Since sinon has consistency problems with mocha and doesn't work in
browser, sinon was hacked somehow.

How sinon was hacked:

The script sinon-4.0.1.js was loaded from

  http://sinonjs.org/releases/sinon-4.0.1.js

After that it was placed to

  /usr/lib/node_modules/sinon

And then the path is written to the html page. So, the full 2Mb script
is loaded everytime.

Also the general test run page test_all.html was made to run all tests
for all JavaScript modules. In such manner it is possible to have
independent tests: function f() in file1.js doesn't overwrite function
f() in file2.js and otherwise.