Replace buttons' states in the forum page
Before, there were two img tags for every button and JavaScript used
style property display for switching them.

Now, every button is an image with two paths to state pictures in
extra data tag attributes - for state on and for state off.
And JavaScript replaces style property src of an image to the right
state.