Fix setting default values in escapeUrlForMessage()
Before this, when user left empty the escape characters field the url
set by user removed and replaced to the default url value. It was
uncomfortable for the user and led to insertion of the user's url
again.

Now, the empty characters field is used while converting the url.

Also if the url field is empty it is replaced to the default url, but
the escape characters field left as is (even if it is empty).